+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ## [Unreleased]
 ### Fixed
 - encoding issue in the product export
+- handling of shopgate special offers
 
 ## [2.9.74] - 2018-01-09
 ### Fixed