+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [2.9.75] - 2018-01-25
 ### Fixed
 - identification of the main image for parent products in export
 - base price in the product export
@@ -709,7 +711,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - migrate plugin from Shopware 3.5.x to 4.0.x
 - use doctrine models
 
-[Unreleased]: https://github.com/shopgate/interface-shopware/compare/2.9.74...HEAD
+[Unreleased]: https://github.com/shopgate/interface-shopware/compare/2.9.75...HEAD
+[2.9.75]: https://github.com/shopgate/interface-shopware/compare/2.9.74...2.9.75
 [2.9.74]: https://github.com/shopgate/interface-shopware/compare/2.9.73...2.9.74
 [2.9.73]: https://github.com/shopgate/interface-shopware/compare/2.9.72...2.9.73
 [2.9.72]: https://github.com/shopgate/interface-shopware/compare/2.9.71...2.9.72